Fix setup.sh for macOS (#24)

+++ setup.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [[ "$OSTYPE" == "darwin"* ]]; then
+   echo "Setting up for MacOS..."
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+   echo "Setting up for Linux..."
+fi
+
 # Node is required for Copilot.
 if [[ "$OSTYPE" == "darwin"* ]]; then
     brew list -q node || brew install node
@@ -15,36 +21,58 @@ elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
 fi
 
 # neovim
-curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim.appimage
-chmod u+x nvim.appimage
-./nvim.appimage --appimage-extract
-./squashfs-root/AppRun --version
-
-sudo mv squashfs-root /
-sudo ln -s /squashfs-root/AppRun /usr/bin/nvim
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    brew install neovim
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim.appimage
+    chmod u+x nvim.appimage
+    ./nvim.appimage --appimage-extract
+    ./squashfs-root/AppRun --version
+    sudo mv squashfs-root /
+    sudo ln -s /squashfs-root/AppRun /usr/bin/nvim
+fi
 
 echo 'alias vi=nvim' >> ~/.bashrc 
 
 # xclip clipboard
-sudo apt-get update
-sudo apt-get install xclip
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    brew install xclip
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    sudo apt-get update
+    sudo apt-get install xclip
+fi
 
 # telescope live_grep
-sudo apt-get install ripgrep
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    brew install ripgrep
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    sudo apt-get install ripgrep
+fi
 
 # lsp
 # python
-sudo apt-get install npm
-sudo npm i -g pyright 
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    brew install npm
+    sudo npm i -g pyright
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    sudo apt-get install npm
+    sudo npm i -g pyright
+fi
 
 # c, cpp, objc, objcpp, cuda, proto
-wget https://apt.llvm.org/llvm.sh
-chmod +x llvm.sh
-sudo ./llvm.sh 16
-sudo update-alternatives --install /usr/bin/clangd clangd /usr/bin/clangd-16 100
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    brew install llvm
+elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    wget https://apt.llvm.org/llvm.sh
+    chmod +x llvm.sh
+    sudo ./llvm.sh 16
+    sudo update-alternatives --install /usr/bin/clangd clangd /usr/bin/clangd-16 100
+fi
 
 # sshfs
-sudo apt-get install sshfs
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    sudo apt-get install sshfs
+fi
 
 # packer
 git clone --depth 1 https://github.com/wbthomason/packer.nvim\
@@ -55,4 +83,6 @@ git clone https://github.com/github/copilot.vim.git \
   ~/.config/nvim/pack/github/start/copilot.vim 2> /dev/null
 
 # cleanup
-rm nvim.appimage llvm.sh
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    rm nvim.appimage llvm.sh
+fi
